backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Environment & Config
python-dotenv==1.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.3.0

# Vector Database
pymilvus==2.4.9

# Embeddings & ML
sentence-transformers==3.3.1

# LLM Integration
groq==0.11.0

# PDF Processing
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==11.0.0

# Object Storage
minio==7.2.11

# Caching
redis==5.0.8
# hiredis==2.3.2  # Optional: Performance optimization (requires C++ build tools on Windows)

# Utilities
python-dateutil==2.9.0
